requirements.txt
torch>=2.3.0
transformers>=4.43.0
accelerate>=0.33.0
peft>=0.11.1
datasets>=2.20.0
sentencepiece>=0.2.0
protobuf>=4.25.3
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.1.0
matplotlib>=3.8.0
python-docx>=1.1.2
docx2txt>=0.8
pypdf>=4.2.0
langdetect>=1.0.9
pyyaml>=6.0.1
tqdm>=4.66.0